feat: Add DashboardSectionsDivider component
